SINGA-37 Enable users to set parameter sharing in model configuration
A share_from field is added into the ParamProto.
It indicates the name of the Param that owns the values of this param.
The NeuralNet class will then share the data_ filed from the owner Param.
In specific, it will consider share_from firstly and then consider param sharing
due to model partition (on dim 0).

Can Compile now but needs test.