SINGA-9 Add Support for Restricted Boltzman Machine (RBM) model
This ticket is to implement RBM in SINGA.
To training RBM models, the Contrastive Divergence (CD) algorithm should be implemented.
We have implemented a BPWorker to run the Back-Propagation algorithm. To implement the CD algorithm, we follow the same way to create a CDWorker whose RunOneBatch function controls the logic of the CD algorithm, including positive phase, negative phase and computing gradient phase.
RBM's layers are different to the layers for feed-forward neural networks, hence new layers for RBM models should be added.