SINGA-21 Code review 4
Update layers for RBM.
The CD algorithm follows Hinton's science paper to do sampling (only
hidden layer is sampled).
May add configuration fields to control the sampling of each layer.
Note. The first gibbs iteration samples the postivie data of the hidden
layer (not the negative data, which is uninitialized).